created gui folder in src and moved gui1.py  into it
